+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## HEAD
+
+* Add `:directory_permissions` to `Storage::FileSystem` (janko-m)
+
 ## 2.2.0 (2016-07-29)
 
 * Soft deprecate `:phase` over `:action` in `context` (janko-m)